fix: JUCE FontOptions, safer casts, and FlexBox integer snapping
Use FontOptions for slider text boxes, tighten float/int conversions in
look-and-feel and FlexBox layout so component bounds snap cleanly without
truncation drift.

+++ components/CustomTabLookAndFeel.h
@@ -104,7 +104,7 @@ public:
     int getTabButtonBestWidth(juce::TabBarButton& button, int tabDepth) override
     {
         juce::GlyphArrangement ga;
-        ga.addLineOfText(getTabButtonFont(button, tabDepth),
+        ga.addLineOfText(getTabButtonFont(button, static_cast<float>(tabDepth)),
                          button.getButtonText(), 0.0f, 0.0f);
         int textWidth = (int) std::ceil(
             ga.getBoundingBox(0, -1, true).getWidth());
